# app/backend/src/main/java/com/example/app/user/web/dto/RegisterRequest.java
package com.example.app.user.web.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record RegisterRequest(
        @Email @NotBlank String email,
        @NotBlank @Size(min = 6, max = 64) String password,
        @NotBlank @Size(max = 100) String displayName
) {
}